requirements.txt
request
beautifulsoup4
openai
python-dotenv
pandas